% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-1-predict.R
\name{cutoff.calc}
\alias{cutoff.calc}
\title{Cutoff Calculation for binary class predictions}
\usage{
cutoff.calc(cutoff, trainObj = NULL, modelObj = NULL)
}
\arguments{
\item{cutoff}{If it's a number between 0 and 1, return this number.
If cutoff == 'prevalence', the prevalence of the first class will be returned.
if cutoff == 'optimal', the cutoff will be optimized by maximizing AUC using the training set.}

\item{trainObj}{A \code{wClsObj} object.}

\item{modelObj}{A \code{wcls} object.}
}
\value{
A \code{numeric} value as the cutoff
}
\description{
Cutoff Calculation for binary class predictions via different ways.
}
