% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-1-grid-search.R
\name{cvGrid}
\alias{cvGrid}
\title{Grid Search via K-fold Cross Validation}
\usage{
cvGrid(trainObj, k.fold = 10, gsCtrl, cutoff = 0.5, criteria = c("acc",
  "sens", "spec", "auc", "ppv", "npv", "mse", "w.acc", "a.acc", "val.f",
  NA)[ifelse("RegObj" \%in\% class(trainObj), 7, 4)], weight = rep(1,
  length(trainObj@X)[1]), inclusion = rep(TRUE, length(trainObj@X)[1]),
  print.fold = FALSE, progress = TRUE, all.models = FALSE)
}
\arguments{
\item{trainObj}{A \code{RegObj} or \code{wClsObj} containing the training set.}

\item{k.fold}{The number of folds in the cross validation.}

\item{gsCtrl}{A \code{list} of arguments that specifies a regression or classification method, handled by
\code{\link{gsCtrlPanel}}.}

\item{cutoff}{A \code{numeric} value used as a cutoff to calculate accuracy when the outcome is bianry.}

\item{criteria}{A criteria to decide for the best tuning parameters. If NA, then the best model will not be picked.}

\item{weight}{A \code{numeric} vector indicating the weight of all samples. Used for the 'w.acc' criteria.}

\item{inclusion}{A \code{logical} vector indicating whether a sample shoule be included in the performance evaluation.}

\item{print.fold}{A \code{logical} value indicating whether each fold should be reported while model fitting.}

\item{progress}{A \code{logical} value indicating whether a progress bar would be used to show the progress.}

\item{all.models}{A \code{logical} value indicating whether to save all fitted models.}
}
\value{
A \code{\link{cvGrid-class}} object.
}
\description{
Grid Search for hyper-parameters in a \code{reg} or \code{wcls} model via K-fold Cross
 Validation.
}
\seealso{
\code{\link{reg}}\cr
\code{\link{wcls}}\cr
\code{\link{ensemble}}\cr
\code{\link{cvKfold}}\cr
\code{\link{workflow}}\cr
}
