% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{cvKfold-class}
\alias{cvKfold-class}
\alias{show,cvKfold-method}
\title{An S4 class containing the K-fold criss validation results}
\usage{
\S4method{show}{cvKfold}(object)
}
\arguments{
\item{object}{A \code{cvKfold} object.}
}
\description{
An S4 class containing the K-fold criss validation results
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{cvKfold} object.
}}

\section{Slots}{

\describe{
\item{\code{pred.table}}{A table for all cross-validation predictions from all models}

\item{\code{prob.table}}{An array with all estimated probability tables}

\item{\code{models}}{The final models returned}

\item{\code{cutoff.list}}{A list of cutoff values}

\item{\code{performances}}{The performance table for all models}

\item{\code{v.score}}{A table of variables scores averaged on all k folds from all models}

\item{\code{pred.error}}{A table of corss-validation prediction error for all models}
}}

\seealso{
\code{\link{ModelObj-class}}\cr
\code{\link{ModelEnsembleObj-class}}\cr
\code{\link{ITRObj-class}}\cr
\code{\link{PredObj-class}}\cr
\code{\link{ListPredObj-class}}\cr
\code{\link{cvGrid-class}}\cr
\code{\link{cvKfold-class}}\cr
\code{\link{workflow-class}}\cr
}
