% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-3-cross-validation.R
\name{cvKfold}
\alias{cvKfold}
\title{K-fold Cross Validation}
\usage{
cvKfold(trainObj, k.fold = 10, cvCtrl.list, cutoff = 0.5,
  unbalance.trt = c("None", "Over", "Under")[1], final.models = FALSE)
}
\arguments{
\item{trainObj}{A \code{RegObj} or \code{wClsObj} containing the training set.}

\item{k.fold}{The number of folds in the cross validation.}

\item{cvCtrl.list}{A \code{list} of \code{list}s of arguments that specifies a regression or classification method.
Each list of arguments is handled by \code{\link{itrCtrlPanel}}.}

\item{cutoff}{Cutoff used when predicting the current fold that not used in training the model, passed to \code{cutoff.calc} function.}

\item{unbalance.trt}{A character string that specifies if sampleing techniques should be applied to
data with unbalanced numbers for each class.}

\item{final.models}{A \code{logical} value indicating whether the final models are kept in the outout.}
}
\value{
A \code{\link{cvKfold-class}} object.
}
\description{
Randomly cut the training set into k-folds and Conduct a K-fold Cross Validation.
 A list of models can be evaluated in the same function. The performance for each model
 will be evaluated as required.
}
\seealso{
\code{\link{cutoff.calc}}\cr
\code{\link{reg}}\cr
\code{\link{wcls}}\cr
\code{\link{cvGrid}}\cr
\code{\link{ensemble}}\cr
\code{\link{workflow}}\cr
}
