% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWNNEL.R
\docType{package}
\name{DWNNEL}
\alias{DWNNEL}
\alias{DWNNEL-package}
\title{DWNNEL: Doubly Weighted Neural Network Ensemble Learning Suite}
\description{
The DWNNEL package provides a doubly weighted neural network ensemble learning
method to model individualized treatment rule by adopting outcome weighted learning,
deep learning and ensemble learning. Besides, a few other frameworks and regression/
classifcation methods are also integrated in this packages, together with cross
validation for either grid search or performance evaluation.
}
