% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2-ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble Learning to extend single model}
\usage{
ensemble(object, n.ensemble = 100, cutoff = 0.5, esCtrl,
  unbalance.trt = c("None", "Over", "Under", "Balance")[1], prop.keep = 1,
  best.opti = TRUE, min.keep = 10, parallel = FALSE, n.nodes = 4)
}
\arguments{
\item{object}{A \code{RegObj} or \code{wClsObj} containing the training set.}

\item{n.ensemble}{The number of models in the ensemble.}

\item{cutoff}{Cutoff used when predicting OOB samples, passed to \code{cutoff.calc} function.}

\item{esCtrl}{A \code{list} of arguments that specifies a regression or classification method, handled by
\code{\link{esCtrlPanel}}.}

\item{unbalance.trt}{A character string that specifies if sampleing techniques should be applied to
data with unbalanced numbers for each class.}

\item{prop.keep}{A \code{numeric} value that specifies the percentage of models to keep with top performance scores.}

\item{best.opti}{A \code{logical} value, whether the number of model to keep is determined automatically.}

\item{min.keep}{A \code{numeric} value that specifies the minimum models to keep.}
}
\value{
A \code{\link{ModelEnsembleObj-class}} object.
}
\description{
Ensemble Learning to extend a \code{reg}, \code{wcls} or a \code{cvGrid} method. Multiple models
 will be fitted using randomly generated bootstrap samples. The out-of-bag (OOB) samples will be used as
 the validation set if the corresponding method could use the validation to fit the model. The
 predictions from all models will be aggregated when predicting for new data. Each model will be
 evaluated and assigned a performance score.
}
\seealso{
\code{\link{cutoff.calc}}\cr
\code{\link{reg}}\cr
\code{\link{wcls}}\cr
\code{\link{cvGrid}}\cr
\code{\link{cvKfold}}\cr
\code{\link{workflow}}\cr
}
