% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-control.R
\name{esCtrlPanel}
\alias{esCtrlPanel}
\title{Manage \code{ensemble} Arguments}
\usage{
esCtrlPanel(machine, ...)
}
\arguments{
\item{machine}{A character string. The \code{reg}, \code{wcls} or \code{cvGrid} function to call.}

\item{...}{Additional arguments passed to the \code{reg}, \code{wcls} or \code{cvGrid} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{ensemble} arguments passed to \code{reg}, \code{wcls} or \code{cvGrid} functions.
}
\seealso{
\code{\link{ensemble}}\cr
}
