% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-1-predict.R
\docType{methods}
\name{eval}
\alias{eval}
\alias{evalPred}
\alias{evalPred,PredObj-method}
\alias{evalPred,ListPredObj-method}
\title{Evaluate the Performance of the Prediction}
\usage{
evalPred(object, test.Y, cutoff = 0.5, type = c("Binary", "Continuous")[1],
  output = ifelse(type == "Continuous", list("mse"), list(c("acc", "sens",
  "spec", "auc", "ppv", "npv", "w.acc", "a.acc", "val.f")))[[1]],
  weight = rep(1, length(test.Y)), inclusion = rep(TRUE, length(test.Y)),
  Y.val = test.Y, Y.resp = rep(1, length(test.Y)))

\S4method{evalPred}{PredObj}(object, test.Y, cutoff = 0.5,
  type = c("Binary", "Continuous")[1], output = ifelse(type == "Continuous",
  list("mse"), list(c("acc", "sens", "spec", "auc", "ppv", "npv", "w.acc",
  "a.acc", "val.f")))[[1]], weight = rep(1, length(test.Y)),
  inclusion = rep(TRUE, length(test.Y)), Y.val = test.Y, Y.resp = rep(1,
  length(test.Y)))

\S4method{evalPred}{ListPredObj}(object, test.Y, cutoff, type, output, weight,
  inclusion, Y.val)
}
\arguments{
\item{object}{A \code{PredObj} or a \code{ListPredObj} object.}

\item{test.Y}{The true response.}

\item{cutoff}{A cutoff for binary response.}

\item{type}{The type of the response, either "Binary" or "Continuous".}

\item{output}{The output types. See details.}

\item{weight}{Sample weight for weighted accuracy.}

\item{inclusion}{Used in thresholded accuracy.}

\item{Y.val}{Used in the value function.}
}
\value{
A \code{data.frame} of performance will be retured.
}
\description{
Evaluate the performance of the prediction using different criteria.
}
\details{
acc: Accuracy
sens: sensitivity
spec: specificity
auc: area under the curve
ppv: positive predictive value
npv: negative predictive value
w.acc: weighted accuracy --- w.acc = sum((Y==pred.Y)*weight)/sum(weight)
a.acc: thresholded accuracy --- a.acc = sum((Y==pred.Y)*inclusion)/sum(inclusion)
val.f: value function  --- val.f = sum((Y.val==pred.Y)*weight)/sum(weight)
}
\seealso{
\code{\link{dwnnel-predict}}
}
