% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\docType{methods}
\name{generateTrtData}
\alias{generateTrtData}
\alias{generateTrtData,TrtDataIdeal-method}
\title{Randomly assgin treatment and generate tratment data object}
\usage{
generateTrtData(object, p.case = 0.5, inc.cutoff = 0, epsilon = 1, ...)

\S4method{generateTrtData}{TrtDataIdeal}(object, p.case = 0.5,
  inc.cutoff = 0, epsilon = 1, ...)
}
\arguments{
\item{object}{A \code{TrtDataIdeal} object.}

\item{p.case}{A number between 0 and 1, indicating the probability to assign the first treatment.}

\item{inc.cutoff}{A positive number, indicating the cutoff of inclusion.}

\item{epsilon}{A random variable with mean zero and variance epsilon^2 will be added to the treatment response.}

\item{...}{Other possible parameters passed to this function.}
}
\value{
An \code{TrtDataObj} object.
}
\description{
From \code{TrtDataIdeal} object, randomly assign treatment labels to all samples and
generate a \code{TrtDataObj} object.
}
\section{Methods (by generic)}{

\code{generateTrtData:} Method to randomly assgin treatment and generate tratment data object
}

\seealso{
\code{\link{TrtDataObj-class}}
\code{\link{TrtDataIdeal-class}}
}
