% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-control.R
\name{gsCtrlPanel}
\alias{gsCtrlPanel}
\title{Manage \code{cvGrid} Arguments}
\usage{
gsCtrlPanel(machine, ...)
}
\arguments{
\item{machine}{A character string. The \code{reg} or \code{wcls} function to call.}

\item{...}{Additional arguments passed to the \code{reg} or \code{wcls} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{cvGrid} arguments passed to \code{reg} or \code{wcls} functions.
}
\seealso{
\code{\link{cvGrid}}\cr
}
