% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{impoorIdeal}
\alias{impoorIdeal}
\title{Import Data to create a \code{TrtDataIdeal} object.}
\usage{
impoorIdeal(X, trtResp1, trtResp2, label)
}
\arguments{
\item{X}{A \code{matrix} or \code{data.frame} containing all samples/variables. It has to be \code{numeric}
and cannot be left blank. Any variable with missing value will be removed.}

\item{trtResp1}{A \code{numeric} vector, indicating the response under the first treatment.}

\item{trtResp2}{A \code{numeric} vector, indicating the response under the second treatment.}

\item{label}{A \code{character} vector of length two, the unique treatment labels. The default is
\code{levels(trtLabl)}.}
}
\value{
An \code{TrtDataIdeal} object.
}
\description{
A user-friendly function to create a \code{TrtDataIdeal} object.
}
\details{
This function takes data set and generate an \code{TrtDataIdeal} in a flexible way:

X can be either a \code{matrix} of a \code{data.frame}, of which each column is a numeric predictor
 of the data set and each row is all predictor values for a sample.
 X should contain all samples/variables.
}
\seealso{
\code{\link{TrtDataIdeal-class}}
}
