% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{importData}
\alias{importData}
\title{Import Data to create a \code{TrtDataObj} object.}
\usage{
importData(X, trtResp = NULL, trtLabl = NULL, trtLevl = levels(trtLabl),
  trtTrue = NULL, sample.weight = rep(1, dim(X)[1]),
  sample.inclsn = rep(TRUE, dim(X)[1]))
}
\arguments{
\item{X}{A \code{matrix} or \code{data.frame} containing all samples/variables. It has to be \code{numeric}
and cannot be left blank. Any variable with missing value will be removed.}

\item{trtResp}{A \code{numeric} vector, indicating the response under assigned treatment. If it's left blank, t
his object could not be used to build models but could be used in \code{predict} function.}

\item{trtLabl}{A \code{factor} vector, the sample treatment labels.}

\item{trtLevl}{A \code{character} vector of length two, the unique treatment labels. The default is
\code{levels(trtLabl)}.}

\item{trtTrue}{A \code{factor} vector, true optimal treatment labels. Left missing if not provided.}

\item{sample.weight}{A \code{numeric} vector, sample weight used to evaluate the performance.}

\item{sample.inclsn}{A \code{logical} vector, indicating whether a sample is used to evaluate the performance.}
}
\value{
An \code{TrtDataObj} object.
}
\description{
A user-friendly function to create a \code{TrtDataObj} object.
}
\details{
This function takes data set and generate an \code{TrtDataObj} in a flexible way:

X can be either a \code{matrix} of a \code{data.frame}, of which each column is a numeric predictor
 of the data set and each row is all predictor values for a sample.
 X should contain all samples/variables.
}
\seealso{
\code{\link{TrtDataObj-class}}
}
