% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4-owl.R
\docType{methods}
\name{itr}
\alias{itr}
\alias{itrDOWL}
\alias{itrOWL}
\alias{itrVT}
\alias{itrSimple}
\alias{EndLot}
\alias{LASSO}
\alias{OWL_LASSO}
\alias{VT_Single}
\alias{VT_Ensemble}
\alias{itrDOWL,TrtDataObj-method}
\alias{itrOWL,TrtDataObj-method}
\alias{itrVT,TrtDataObj-method}
\alias{itrSimple,TrtDataObj-method}
\title{Individualized Treatment Rule Frameworks}
\usage{
itrDOWL(object, valid = NULL, regCtrl, wclsCtrl, ...)

itrOWL(object, valid = NULL, wclsCtrl, ...)

itrVT(object, valid = NULL, regCtrl, ...)

itrSimple(object, valid = NULL, regCtrl, ...)

EndLot(object, valid = NULL, ...)

LASSO(object, valid = NULL, ...)

OWL_LASSO(object, valid = NULL, ...)

VT_Single(object, valid = NULL, ...)

VT_Ensemble(object, valid = NULL, ...)

\S4method{itrDOWL}{TrtDataObj}(object, valid = NULL, regCtrl, wclsCtrl, ...)

\S4method{itrOWL}{TrtDataObj}(object, valid = NULL, wclsCtrl, ...)

\S4method{itrVT}{TrtDataObj}(object, valid = NULL, regCtrl, ...)

\S4method{itrSimple}{TrtDataObj}(object, valid = NULL, regCtrl, ...)
}
\arguments{
\item{object}{A \code{TrtDataObj} object, used as the training set.}

\item{valid}{A \code{TrtDataObj} object, used as the validation set. The default is \code{NULL}.}

\item{regCtrl}{A \code{list} of arguments that specifies a regression method, handled by
\code{\link{regCtrlPanel}}.}

\item{wclsCtrl}{A \code{list} of arguments that specifies a weighted classification method, handled by
\code{\link{wclsCtrlPanel}}.}

\item{...}{Other parameters passed to \code{itr} functions.}
}
\value{
Returns a \code{ITRObj} object.
}
\description{
A collection of Individualized Treatment Rule frameworks.
}
\details{
These \code{itr} frameworks all works for the personalized treatment problems with
 two treatmnet labels. Different frameworks would call either regression methods
 or classification methods, or both. Methods of the same type could be applied in the
 same framework.
}
\section{Methods (by generic)}{

\code{itrDOWL}: Doubly Weighted Outcome Learning Framework. This procedure first calls a
 regression model to estimate E(Y|X), the expected treatment effect averaged on all
 treatments. Then it calls a weighted classification to estimate the ITR.


\code{itrOWL}: Outcome Weighted Learning Framework. This procedure directly calls a weighted classification
 to estimate the ITR by outcome weighted learning.


\code{itrVT}: Virture Twins Framework. This procedure calls the same method to fit two regression
 models for patients with either treatment. The ITR is determined by comparing the predictions
 with both models.


\code{itrSimple}: Simple Regression Framework. This procedure fits one regression model with
 A*Y ~ X, where A from {1, -1} is the treatment label, Y is the continuous outcome and X is the
 predictors.
}

\references{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3636816/pdf/nihms379905.pdf}{Outcome Weighted Learning}\cr
\href{http://onlinelibrary.wiley.com/doi/10.1002/sim.4322/epdf}{Virtual Twins}\cr
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4338439/pdf/nihms637476.pdf}{A simple Regression}\cr
}
\seealso{
\code{\link{TrtDataObj-class}}\cr
\code{\link{ITRObj-class}}\cr
\code{\link{regCtrlPanel}}\cr
\code{\link{wclsCtrlPanel}}\cr
\code{\link{reg}}\cr
\code{\link{wcls}}\cr
\code{\link{cvGrid}}\cr
\code{\link{ensemble}}\cr
\code{\link{cvKfold}}\cr
\code{\link{workflow}}\cr
}
