% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-control.R
\name{itrCtrlPanel}
\alias{itrCtrlPanel}
\title{Manage \code{cvKfold} or \code{workflow} Arguments}
\usage{
itrCtrlPanel(ITR, ...)
}
\arguments{
\item{ITR}{A character string, the \code{itr} to call.}

\item{...}{Additional arguments passed to the \code{itr} function function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{cvKfold} or \code{workflow} arguments passed to \code{itr} functions.
}
\seealso{
\code{\link{cvKfold}}\cr
\code{\link{workflow}}\cr
}
