\name{ITRLearn-package}
\alias{ITRLearn-package}
\alias{ITRLearn}
\docType{package}
\title{
	Statistical Learing for Individualized Treatment Regime
}
\description{
	Maximin-projection learning (MPL, Shi, et al., 2018) is implemented for 
	recommending a meaningful and reliable individualized treatment regime for future 
	groups of patients based on the observed data from different populations with
	heterogeneity in individualized decision making. Q-learning and A-learning are
	implemented for estimating the groupwise contrast function that shares the same
	marginal treatment effects. The packages contains classical Q-learning and A-learning
	algorithms for a single stage study as a byproduct. 
}
\details{
\tabular{ll}{
Package: \tab ITRLearn\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2018-11-14\cr
License: \tab GPL-2\cr
}

}
\author{
Chengchun Shi, Rui Song, Wenbin Lu and Bo Fu

Maintainer: Chengchun Shi <cshi4@ncsu.edu>
}
\references{
Shi, C., Song, R., Lu, W., and Fu, B. (2018). Maximin Projection Learning for Optimal Treatment 
Decision with Heterogeneous Individualized Treatment Effects. \emph{Journal of the Royal Statistical
Society, Series B,} \bold{ 80:} 681-702.
}