% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{ITRObj-class}
\alias{ITRObj-class}
\alias{show,ITRObj-method}
\title{An S4 class, an individualized treatment rule object}
\usage{
\S4method{show}{ITRObj}(object)
}
\arguments{
\item{object}{A \code{ITRObj} object.}
}
\description{
An S4 class, an individualized treatment rule object
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ITRObj} object.
}}

\section{Slots}{

\describe{
\item{\code{label}}{A character vector of length two, treatment labels}

\item{\code{model}}{A list of regression or classification models related to the indivifualized treatment rule.}

\item{\code{model.type}}{Type of individualized treatment rule, including \code{itrDOWL}, \code{itrOWL}, \code{itrVT}, \code{itrSimple}}
}}

\seealso{
\code{\link{TrtDataObj-class}}\cr
\code{\link{itrDOWL}}\cr
\code{\link{itrOWL}}\cr
\code{\link{itrVT}}\cr
\code{\link{itrSimple}}\cr
}
