% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R
\docType{class}
\name{ListPredObj-class}
\alias{ListPredObj-class}
\title{An S4 class containing the prediction}
\description{
An S4 class containing the prediction
}
\section{Slots}{

\describe{
\item{\code{listPred}}{A list of \code{PredObj} objects.}
}}

\seealso{
\code{\link{PredObj-class}}\cr
}
