% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{ModelEnsembleObj-class}
\alias{ModelEnsembleObj-class}
\alias{show,ModelEnsembleObj-method}
\title{An S4 class, an ensemble model object}
\usage{
\S4method{show}{ModelEnsembleObj}(object)
}
\arguments{
\item{object}{A \code{ModelEnsembleObj} object.}
}
\description{
An S4 class, an ensemble model object
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ModelEnsembleObj} object.
}}

\section{Slots}{

\describe{
\item{\code{model.list}}{A list of regression or classification models in the ensemble}

\item{\code{model.type}}{Type of the model, either "Continuous" or "Binary"}

\item{\code{loss}}{A numeric vector, the loss of each model}

\item{\code{keep}}{A numeric vector or matrix, indicating whether each model is included when aggregating the predictions}
}}

\seealso{
\code{\link{wClsObj-class}}\cr
\code{\link{RegObj-class}}\cr
\code{\link{ModelObj-class}}\cr
}
