% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{ModelObj-class}
\alias{ModelObj-class}
\alias{show,ModelObj-method}
\title{An S4 class, a regression or classification model object}
\usage{
\S4method{show}{ModelObj}(object)
}
\arguments{
\item{object}{A \code{ModelObj} object.}
}
\description{
An S4 class, a regression or classification model object
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ModelObj} object.
}}

\section{Slots}{

\describe{
\item{\code{model}}{A regression or classification model}

\item{\code{model.type}}{Type of the model, either "Continuous" or "Binary"}
}}

\seealso{
\code{\link{wClsObj-class}}\cr
\code{\link{RegObj-class}}\cr
\code{\link{wclskSVM}}\cr
\code{\link{wclslSVM}}\cr
\code{\link{wclsLASSO}}\cr
\code{\link{wclsANN}}\cr
\code{\link{wclsDNN}}\cr
\code{\link{regLASSO}}\cr
\code{\link{regSVR}}\cr
\code{\link{regRF}}\cr
\code{\link{regANN}}\cr
\code{\link{regDNN}}\cr
\code{\link{ModelEnsembleObj-class}}\cr
}
