% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R
\docType{class}
\name{PredContinuousObj-class}
\alias{PredContinuousObj-class}
\title{An S4 class containing the prediction}
\description{
An \code{PredObj-class} sub-class for model dealing with a continuous response.
}
