% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R
\docType{class}
\name{PredDiscreteObj-class}
\alias{PredDiscreteObj-class}
\title{An S4 class containing the prediction}
\description{
An \code{PredObj-class} sub-class for model dealing with a discrete response.
}
\section{Slots}{

\describe{
\item{\code{prob}}{The matrix for estimated probability for each class.}
}}

