% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{PredObj-class}
\alias{PredObj-class}
\alias{show,PredObj-method}
\title{An S4 class containing the prediction}
\usage{
\S4method{show}{PredObj}(object)
}
\arguments{
\item{object}{A \code{PredObj} object.}
}
\description{
An S4 class containing the prediction
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{PredObj} object.
}}

\section{Slots}{

\describe{
\item{\code{pred}}{Prediction of new data from some model}

\item{\code{v.score}}{Potential variable importance score, not appicable in this version}
}}

\seealso{
\code{\link{ModelObj-class}}\cr
\code{\link{ModelEnsembleObj-class}}\cr
\code{\link{ITRObj-class}}\cr
\code{\link{cvGrid-class}}\cr
\code{\link{cvKfold-class}}\cr
\code{\link{workflow-class}}\cr
}
