% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1-regression.R
\docType{methods}
\name{reg}
\alias{reg}
\alias{regLASSO}
\alias{regSVR}
\alias{regRF}
\alias{regANN}
\alias{regDNN}
\alias{regLASSO,RegObj-method}
\alias{regSVR,RegObj-method}
\alias{regRF,RegObj-method}
\alias{regANN,RegObj-method}
\alias{regDNN,RegObj-method}
\title{Regression Methods}
\usage{
regLASSO(object, valid = NULL, ...)

regSVR(object, valid = NULL, ...)

regRF(object, valid = NULL, ...)

regANN(object, valid = NULL, ...)

regDNN(object, valid = NULL, ...)

\S4method{regLASSO}{RegObj}(object, valid = NULL, ...)

\S4method{regSVR}{RegObj}(object, valid = NULL, ...)

\S4method{regRF}{RegObj}(object, valid = NULL, ...)

\S4method{regANN}{RegObj}(object, valid = NULL, ...)

\S4method{regDNN}{RegObj}(object, valid = NULL, ...)
}
\arguments{
\item{object}{A \code{regObj} object, used as a training set.}

\item{valid}{A \code{wClsObj} object, used as a validation set.}

\item{...}{Arguments passed to the function being called.}
}
\value{
Returns a \code{ModelObj} object.
}
\description{
A collection of functions to build regression models.
}
\details{
These \code{reg} methods construct a single regression model given a \code{regObj}
 and a set of parameters. This function returns a \code{ModelObj}.
}
\section{Methods (by generic)}{

\code{regLASSO:} Method to build a regression model using glmnet::glmnet.


\code{regSVR:} Method to build a regression model using e1071::svm.


\code{regRF:} Method to build a regression model using randomForest::randomForest.


\code{regANN:} Method to build a regression model using nnet::nnet.


\code{regDNN:} Method to build a deep neural network regression model using nn.regresser function.
}

\seealso{
\code{\link{reg}}\cr
\code{\link{wcls}}\cr
\code{\link{cvGrid}}\cr
\code{\link{cvKfold}}\cr
\code{\link{workflow}}\cr
}
