% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1-regression.R
\name{REG_}
\alias{REG_}
\title{Workhorse for reg Methods}
\usage{
REG_(object, valid, regMethod, ...)
}
\arguments{
\item{object}{A \code{regObj} object, used as a training set.}

\item{valid}{A \code{wClsObj} object, used as a validation set.}

\item{regMethod}{A function call unique to that reg method.}

\item{...}{Arguments passed to the function being called.}
}
\value{
Returns an \code{ModelObj} object.
}
\description{
Used as a back-end wrapper for creating new reg methods.
}
