% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-control.R
\name{regCtrlPanel}
\alias{regCtrlPanel}
\title{Manage \code{itr} Arguments}
\usage{
regCtrlPanel(machine, ...)
}
\arguments{
\item{machine}{A character string, the \code{reg} function, \code{ensemble} or \code{cvGrid} to call.}

\item{...}{Additional arguments passed to the \code{reg} function, \code{ensemble} or \code{cvGrid} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{itr} arguments passed to \code{reg} functions, \code{cvGrid}
 or \code{ensemble}.
}
\seealso{
\code{\link{itr}}\cr
}
