% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{RegObj-class}
\alias{RegObj-class}
\alias{show,RegObj-method}
\title{An S4 class, a data object to conduct a regression.}
\usage{
\S4method{show}{RegObj}(object)
}
\arguments{
\item{object}{A \code{RegObj} object.}
}
\description{
An S4 class, a data object to conduct a regression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{RegObj} object.
}}

\section{Slots}{

\describe{
\item{\code{Y}}{A numeric vector, the response}

\item{\code{X}}{A numeric matrix. Containing all the variables.}
}}

\seealso{
\code{\link{regLASSO}}\cr
\code{\link{regSVR}}\cr
\code{\link{regRF}}\cr
\code{\link{regANN}}\cr
\code{\link{regDNN}}\cr
}
