% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-method-def.R
\name{show.df}
\alias{show.df}
\title{Print first several columns of a data.frame in a slot of an object.}
\usage{
show.df(object, slot.name, k = ncol(methods::slot(object, slot.name)),
  l = nrow(methods::slot(object, slot.name)))
}
\arguments{
\item{object}{The object.}

\item{slot.name}{The slot name storing the data.frame}

\item{k}{First k columns will be printed.}

\item{l}{First l rows will be printed.}
}
\description{
\code{show} method call this function to print a data.frame.
}
