% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-method-def.R
\name{show.vec}
\alias{show.vec}
\title{Print first several elements of a vector in a slot of an object.}
\usage{
show.vec(object, slot.name, k = 10)
}
\arguments{
\item{object}{The object.}

\item{slot.name}{The slot name storing the vector.}

\item{k}{First k elements will be printed.}
}
\description{
\code{show} method call this function to print a vector.
}
