% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2-split.R
\docType{methods}
\name{splitData}
\alias{splitData}
\alias{splitData}
\alias{splitData,TrtDataObj-method}
\alias{splitData,TrtDataIdeal-method}
\alias{splitData,wClsObj-method}
\alias{splitData,RegObj-method}
\title{Split Data}
\usage{
splitData(object, split)

\S4method{splitData}{TrtDataObj}(object, split)

\S4method{splitData}{TrtDataIdeal}(object, split)

\S4method{splitData}{wClsObj}(object, split)

\S4method{splitData}{RegObj}(object, split)
}
\arguments{
\item{object}{An object to be split.}

\item{split}{A character, numeric variable or a numeric vector declaring a way to split
the object. If it's number between 0 and 1, all samples will be split into two subsets
randomly, with the trainObj containing such proportion of all samples abd validObj containing
the rest. If split is a character and is "bootstrap", the trainObj will be a bootstrap sample
of the original data set and the validObj will contain out-of-bag samples. If split is a vector
of integers, the trainObj will contain samples whose indice are in the vector, and validObj will
contain the rest.}
}
\value{
Returns a list of two objects of the input type, trainObj and validObj, and the indice of the trainObj.
}
\description{
Split an object into a training set and a validation set.
}
\details{
A function to split the objects into a list of two objects of the same type:
 one named trainObj and the other named validObj.
}
\seealso{
\code{\link{TrtDataObj-class}}
\code{\link{TrtDataIdeal-class}}
\code{\link{wClsObj-class}}
\code{\link{RegObj-class}}
}
