% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{TrtDataIdeal-class}
\alias{TrtDataIdeal-class}
\alias{show,TrtDataIdeal-method}
\title{An S4 class, an ideal dataset that contains covariates and responses under both treatments.
It can be used to generate a \code{TrtDataObj} randomly.}
\usage{
\S4method{show}{TrtDataIdeal}(object)
}
\arguments{
\item{object}{A \code{TrtDataIdeal} object.}
}
\description{
An S4 class, an ideal dataset that contains covariates and responses under both treatments.
It can be used to generate a \code{TrtDataObj} randomly.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{TrtDataIdeal} object.
}}

\section{Slots}{

\describe{
\item{\code{trtResp1}}{A numeric vector, the response under the first treatment}

\item{\code{trtResp2}}{A numeric vector, the response under the second treatment}

\item{\code{X}}{A numeric matrix. Containing all the variables.}

\item{\code{label}}{A character vector of length two, two treatment labels.
The first one corresponds to \code{trtResp1} and the second one corresponds to \code{trtResp2}.}
}}

\seealso{
\code{\link{TrtDataIdeal-class}}\cr
\code{\link{TrtDataObj-class}}\cr
\code{\link{generateTrtData}}\cr
}
