% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{TrtDataObj-class}
\alias{TrtDataObj-class}
\alias{show,TrtDataObj-method}
\title{An S4 class containing treatment, covariates and response}
\usage{
\S4method{show}{TrtDataObj}(object)
}
\arguments{
\item{object}{A \code{TrtDataObj} object.}
}
\description{
An S4 class containing treatment, covariates and response
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{TrtDataObj} object.
}}

\section{Slots}{

\describe{
\item{\code{trtResp}}{A numeric vector, the response under assigned treatment}

\item{\code{trtLabl}}{A factor vector, the assigned treatment label}

\item{\code{X}}{A numeric matrix. Containing all the variables.}

\item{\code{trtLevl}}{A character vector of length two, two treatment labels}

\item{\code{trtTrue}}{A factor vector, true the better treatment}

\item{\code{sample.weight}}{A numeric vector, indicating the weight of a sample in performance evaluation}

\item{\code{sample.inclsn}}{A logical vector, indicating whether a sample is included in performance evaluation}
}}

\seealso{
\code{\link{TrtDataObj-class}}\cr
\code{\link{TrtDataIdeal-class}}\cr
\code{\link{wClsObj-class}}\cr
\code{\link{RegObj-class}}\cr
\code{\link{ModelObj-class}}\cr
\code{\link{ModelEnsembleObj-class}}\cr
\code{\link{cvGrid-class}}\cr
\code{\link{cvKfold-class}}\cr
\code{\link{workflow-class}}\cr
}
