% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4-w-svm.R
\name{w.k.svm}
\alias{w.k.svm}
\title{Weighted Kernel SVM classification Model}
\usage{
w.k.svm(Y, X, W = NULL, C = 1, kernel = "gaussian", gamma = 1/dim(X)[2],
  c = 0, d = 3)
}
\arguments{
\item{Y}{A \code{factor} vector with two levels, a binary response.}

\item{X}{A \code{numeric} matrix, the predictors.}

\item{W}{A \code{numeric} matrix, the weights for all samples. The default is 1.}

\item{C}{The coefficient for the regularizer.}

\item{kernel}{The kernel function.
gaussian: exp(-gamma*(u-v)^2)
polynomial: (gamma*u*v + c)^d
sigmoid: tanh(gamma*u*v + c)}

\item{gamma}{Parameter needed for all kernels.}

\item{c}{Parameter needed for polynomial and sigmoid kernels.}

\item{d}{Parameter needed for sigmoid kernel.}
}
\value{
Returns a \code{DnnModelObj} object.
}
\description{
Fit a Weighted Kernel SVM classification Model
}
\seealso{
\code{\link{DnnModelObj-class}}\cr
}
