% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4-w-svm.R
\name{w.l.svm}
\alias{w.l.svm}
\title{Weighted Linear SVM classification Model}
\usage{
w.l.svm(Y, X, W = NULL, C = 1, epsilon = 10^-10)
}
\arguments{
\item{Y}{A \code{factor} vector with two levels, a binary response.}

\item{X}{A \code{numeric} matrix, the predictors.}

\item{W}{A \code{numeric} matrix, the weights for all samples. The default is 1.}

\item{C}{The coefficient for the regularizer.}

\item{epsilon}{A minor value added to make the quadratic system solvable.}
}
\value{
Returns a \code{DnnModelObj} object.
}
\description{
Fit a Weighted Linear SVM classification Model
}
\seealso{
\code{\link{DnnModelObj-class}}\cr
}
