% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-2-w-classification.R
\docType{methods}
\name{wcls}
\alias{wcls}
\alias{wclskSVM}
\alias{wclslSVM}
\alias{wclsLASSO}
\alias{wclsANN}
\alias{wclsDNN}
\alias{wclskSVM,wClsObj-method}
\alias{wclslSVM,wClsObj-method}
\alias{wclsLASSO,wClsObj-method}
\alias{wclsANN,wClsObj-method}
\alias{wclsDNN,wClsObj-method}
\title{Weighted Classification Methods}
\usage{
wclskSVM(object, valid = NULL, ...)

wclslSVM(object, valid = NULL, ...)

wclsLASSO(object, valid = NULL, ...)

wclsANN(object, valid = NULL, ...)

wclsDNN(object, valid = NULL, ...)

\S4method{wclskSVM}{wClsObj}(object, valid = NULL, ...)

\S4method{wclslSVM}{wClsObj}(object, valid = NULL, ...)

\S4method{wclsLASSO}{wClsObj}(object, valid = NULL, ...)

\S4method{wclsANN}{wClsObj}(object, valid = NULL, ...)

\S4method{wclsDNN}{wClsObj}(object, valid = NULL, ...)
}
\arguments{
\item{object}{A \code{wClsObj} object, used as a training set.}

\item{valid}{A \code{wClsObj} object, used as a validation set.}

\item{...}{Arguments passed to the function being called.}
}
\value{
Returns a \code{ModelObj} object.
}
\description{
A collection of functions to build weighted classification models.
}
\details{
These \code{wcls} methods construct a single weighted classification model given a \code{wClsObj}
 and a set of parameters. This function returns a \code{ModelObj}.
}
\section{Methods (by generic)}{

\code{wclskSVM:} Method to build a (non-linear) kernel SVM weighted classification model using w.k.svm function.


\code{wclslSVM:} Method to build a linear SVM weighted classification model using w.l.svm function.


\code{wclsLASSO:} Method to build a weighted classification model using glmnet::glmnet.


\code{wclsANN:} Method to build a weighted classification model using nnet::nnet.


\code{wclsDNN:} Method to build a deep neural network weighted classification model using nn.classifier function.
}

\seealso{
\code{\link{reg}}\cr
\code{\link{wcls}}\cr
\code{\link{cvGrid}}\cr
\code{\link{cvKfold}}\cr
\code{\link{workflow}}\cr

 @examples
}
