% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-2-w-classification.R
\name{wCLS_}
\alias{wCLS_}
\title{Workhorse for wcls Methods}
\usage{
wCLS_(object, valid, clsMethod, ...)
}
\arguments{
\item{object}{A \code{wClsObj} object, used as a training set.}

\item{valid}{A \code{wClsObj} object, used as a validation set.}

\item{clsMethod}{A function call unique to that wcls method.}

\item{...}{Arguments passed to the function being called.}
}
\value{
Returns an \code{ModelObj} object.
}
\description{
Used as a back-end wrapper for creating new wcls methods.
}
