% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-control.R
\name{wclsCtrlPanel}
\alias{wclsCtrlPanel}
\title{Manage \code{itr} Arguments}
\usage{
wclsCtrlPanel(machine, ...)
}
\arguments{
\item{machine}{A character string, the \code{wcls} function, \code{ensemble} or \code{cvGrid} to call.}

\item{...}{Additional arguments passed to the \code{wcls} function, \code{ensemble} or \code{cvGrid} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{itr} arguments passed to \code{wcls} functions, \code{cvGrid}
 or \code{ensemble}.
}
\seealso{
\code{\link{itr}}\cr
}
