% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{wClsObj-class}
\alias{wClsObj-class}
\alias{show,wClsObj-method}
\title{An S4 class, a data object to conduct a weighted classification.}
\usage{
\S4method{show}{wClsObj}(object)
}
\arguments{
\item{object}{A \code{wClsObj} object.}
}
\description{
An S4 class, a data object to conduct a weighted classification.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{wClsObj} object.
}}

\section{Slots}{

\describe{
\item{\code{weight}}{A numeric vector, the sample weight}

\item{\code{Y}}{A factor vector, the response label}

\item{\code{X}}{A numeric matrix. Containing all the variables.}

\item{\code{Y.lvl}}{A character vector of length two, two response labels.}
}}

\seealso{
\code{\link{wclskSVM}}\cr
\code{\link{wclslSVM}}\cr
\code{\link{wclsLASSO}}\cr
\code{\link{wclsANN}}\cr
\code{\link{wclsDNN}}\cr
}
