% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4-w-svm.R
\docType{class}
\name{wkSVMObj-class}
\alias{wkSVMObj-class}
\title{An S4 class containing a weighted kernel SVM classfication model}
\description{
An S4 class containing a weighted kernel SVM classfication model
}
\section{Slots}{

\describe{
\item{\code{alpha}}{Parameter alpha in the fitted model.}

\item{\code{bias}}{The bias term in the kernel SVM model.}

\item{\code{fitted}}{A numeric number, indicating the loss on the validation set}

\item{\code{kernel.f}}{A character string that specifies the kernel function.}

\item{\code{X}}{Predictors from the training set.}

\item{\code{Y}}{Response from the training set.}

\item{\code{support}}{Support vectors from the training set.}

\item{\code{n.support}}{Number of support vectors.}

\item{\code{label}}{A character vector of length two, binary class labels.}
}}

\seealso{
\code{\link{wClsObj-class}}\cr
\code{\link{RegObj-class}}\cr
\code{\link{ModelObj-class}}\cr
\code{\link{ModelEnsembleObj-class}}\cr
\code{\link{cvGrid-class}}\cr
\code{\link{cvKfold-class}}\cr
\code{\link{workflow-class}}\cr
}
