% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4-w-svm.R
\docType{class}
\name{wlSVMObj-class}
\alias{wlSVMObj-class}
\title{An S4 class containing a weighted linear SVM classfication model}
\description{
An S4 class containing a weighted linear SVM classfication model
}
\section{Slots}{

\describe{
\item{\code{beta}}{Coefficients in the fitted linear SVM model.}

\item{\code{fitted}}{The bias term in the kernel SVM model.}

\item{\code{label}}{A character vector of length two, binary class labels.}
}}

\seealso{
\code{\link{wClsObj-class}}\cr
\code{\link{RegObj-class}}\cr
\code{\link{ModelObj-class}}\cr
\code{\link{ModelEnsembleObj-class}}\cr
\code{\link{cvGrid-class}}\cr
\code{\link{cvKfold-class}}\cr
\code{\link{workflow-class}}\cr
}
