% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-class-def.R, R/1-2-method-def.R
\docType{class}
\name{workflow-class}
\alias{workflow-class}
\alias{show,workflow-method}
\title{An S4 class containing the workflow results}
\usage{
\S4method{show}{workflow}(object)
}
\arguments{
\item{object}{A \code{workflow} object.}
}
\description{
An S4 class containing the workflow results
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{workflow} object.
}}

\section{Slots}{

\describe{
\item{\code{pred.array}}{An array for all iterations of all cross-validation predictions from all models}

\item{\code{prob.array}}{An array with all estimated probability tables}

\item{\code{performance.array}}{An array storing the performance score for all iterations}

\item{\code{performances}}{The aggregated performance table for all models}

\item{\code{cutoff.list}}{A list of cutoff values}

\item{\code{v.score}}{A table of variables scores averaged on all k folds from all models}

\item{\code{pred.error}}{A table of corss-validation prediction error for all models}

\item{\code{models}}{The final models returned}
}}

\seealso{
\code{\link{ModelObj-class}}\cr
\code{\link{ModelEnsembleObj-class}}\cr
\code{\link{ITRObj-class}}\cr
\code{\link{PredObj-class}}\cr
\code{\link{ListPredObj-class}}\cr
\code{\link{cvGrid-class}}\cr
\code{\link{cvKfold-class}}\cr
\code{\link{workflow-class}}\cr
}
