% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-1-workflow.R
\name{workflow}
\alias{workflow}
\title{A workflow to Evaluate the Performance of a List of Methods}
\usage{
workflow(trainObj, validObj = NULL, k.fold = 10, n.iter = 10, cvCtrl.list,
  unbalance.trt = c("None", "over", "under")[1], cutoff = 0.5)
}
\arguments{
\item{trainObj}{A \code{RegObj} or \code{wClsObj} containing the training set.}

\item{validObj}{A \code{RegObj} or \code{wClsObj} containing the validation set.}

\item{k.fold}{The number of folds in the cross validation.}

\item{n.iter}{The number of interations of k-fold cross validations.}

\item{cvCtrl.list}{A \code{list} of \code{list}s of arguments that specifies a regression or classification method.
Each list of arguments is handled by \code{\link{itrCtrlPanel}}.}

\item{unbalance.trt}{A character string that specifies if sampleing techniques should be applied to
data with unbalanced numbers for each class.}

\item{cutoff}{Cutoff used when predicting the current fold that not used in training the model, passed to \code{cutoff.calc} function.}
}
\value{
A \code{\link{workflow-class}} object.
}
\description{
In this procedure, if a validation set is provided, all methods in the list will be
 implemented once and the performance is evaluated using the validation set.
}
\details{
If the validation set is missing, an n iterations of k-fold cross validation will
 be performed to evaluate each method.

The final performance scores will be reported. And all final models will be saved.
}
\seealso{
\code{\link{cutoff.calc}}\cr
\code{\link{reg}}\cr
\code{\link{wcls}}\cr
\code{\link{cvGrid}}\cr
\code{\link{ensemble}}\cr
\code{\link{cvKfold}}\cr
}
